+++ src/AdminPackage.php
@@ -44,6 +44,7 @@ final class AdminPackage extends Package
         return [
             'resources/admin.css' => 'css/admin.css',
             'resources/admin.tables.js' => 'js/admin.tables.js',
+            'resources/admin.modal.js' => 'js/admin.modal.js',
         ];
     }